# doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "单点登录系统api文档",
        "title": "单点登录系统",
        "contact": {
            "name": "wuzehui"
        },
        "version": "1.0"
    },
    "host": "sso.com:8099",
    "basePath": "/",
    "paths": {
...
        "/api/v1/clients": {
            "get": {
                "description": "客户端列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端管理"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/static.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Client"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "添加客户端",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "回调地址",
                        "name": "callback",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "域名",
                        "name": "domain",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用来签名校验的密钥",
                        "name": "secret",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/clients/{id}": {
            "put": {
                "description": "客户端更新",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端管理"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "回调地址",
                        "name": "callback",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "域名",
                        "name": "domain",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用来签名校验的密钥",
                        "name": "secret",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    }
                }
            }
        },
...
        "/svc/session": {
            "post": {
                "description": "客户端回调确认接口，更新session状态为已登录",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Svc接口"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端域名",
                        "name": "domain",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "签名",
                        "name": "sign",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "时间戳",
                        "name": "timestamp",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    }
                }
            }
        },
        "/svc/userinfo": {
            "post": {
                "description": "客户端业务代码执行之前，调用该接口获取用户信息",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Svc接口"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户端域名",
                        "name": "domain",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "签名",
                        "name": "sign",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "时间戳",
                        "name": "timestamp",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/static.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Client": {
            "type": "object",
            "properties": {
                "callback": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
...
        "static.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}